calculator.js is imported into Calculator.jsx
