--- src/components/useHandleButtonOperation.jsx
@@ -1,14 +1,0 @@
-import { useState } from 'react';
-import calculate from '../logic/calculate';
-
-function useHandleButtonOperation() {
-  const [currentState, setCurrentState] = useState({ total: 0, next: null, operation: null });
-
-  const addDigit = (buttonName) => {
-    setCurrentState(calculate(currentState, buttonName));
-  };
-
-  return [currentState, addDigit];
-}
-
-export default useHandleButtonOperation;
